+++ CHANGELOG.md
@@ -16,6 +16,11 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### New features
 
+* `jj desc` adds the commit trailers, configured in the `commit_trailers` template, to the commit description. Use
+  cases include DCO Sign Off and Gerrit Change Id.
+
+* `format_signed_off_by_trailer(commit)` and `format_gerrit_change_id_trailer(commit)` commit templates has been added.
+
 ### Fixed bugs
 
 ## [0.28.1] - 2025-04-04